fixed a few windows warnings

+++ src/mains/hello_triangle/hello_triangle.cpp
@@ -166,7 +166,8 @@ int main()
         timeValue = glfwGetTime();
         greenValue = sin(timeValue) / 2.0f + 0.5f;
         vertexColorLocation = glGetUniformLocation(shaderProgram, "ourColor");
-        glUniform4f(vertexColorLocation, 0.0f, greenValue, 1.0f - greenValue, 1.0f);
+        glUniform4f(vertexColorLocation, 0.0f, static_cast<GLfloat>(greenValue),
+            static_cast<GLfloat>(1.0f - greenValue), 1.0f);
 
         // as we only have a single VAO there's no need to bind it every time,
         // but we'll do so to keep things a bit more organized
